tools/requirements.txt
pillow
numpy
scipy
tqdm
mss
pyautogui
pytesseract
opencv-python
pyperclip